[Compiler]  Further Split long expression files (#884)
* add comments to expression

* change namespace style

* remove

* update neo to latest

* adapt to latest neo version

* dotnet format

* Apply suggestions from code review

Clean empty lines

* remove comments

* update neo

* Update BinaryExpression.cs

* Update src/Neo.Compiler.CSharp/MethodConvert/Expression/CastExpression.cs

* Update src/Neo.Compiler.CSharp/MethodConvert/Expression/CastExpression.cs

* merge master

---------
